Fixed broken links between pages
Deleted the / from the links between Orange Chicken, Kung Pao, and
Chow Mein as they are in the same directory and do not need the /s.

Also, fixed the link back to home as it only had one period but due to
the recipes being in a directory the link needed to leave that directory
and get back to the root directory so it needed double period.